task1_rfq_crm/requirements.txt
python-dotenv==1.0.1
pytest==8.3.2
# Optional — only needed for live mode
openai>=1.30.0
boto3>=1.34.0
